Change port name to cxd56
