+++ conf.py
@@ -125,6 +125,7 @@ exclude_patterns = ["**/build*",
                     "ports/cc3200",
                     "ports/cc3200/FreeRTOS",
                     "ports/cc3200/hal",
+                    "ports/cxd56/spresense-exported-sdk",
                     "ports/esp32",
                     "ports/esp8266/boards",
                     "ports/esp8266/common-hal",
@@ -142,7 +143,6 @@ exclude_patterns = ["**/build*",
                     "ports/stm32f4/README.md",
                     "ports/pic16bit",
                     "ports/qemu-arm",
-                    "ports/spresense/spresense-exported-sdk",
                     "ports/stm32",
                     "ports/stm32/hal",
                     "ports/stm32/cmsis",
