+++ ports/cxd56/common-hal/busio/SPI.h
@@ -23,9 +23,27 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-#ifndef MICROPY_INCLUDED_SPRESENSE_INTERNAL_FLASH_ROOT_POINTERS_H
-#define MICROPY_INCLUDED_SPRESENSE_INTERNAL_FLASH_ROOT_POINTERS_H
 
-#define FLASH_ROOT_POINTERS
+#ifndef MICROPY_INCLUDED_CXD56_COMMON_HAL_BUSIO_SPI_H
+#define MICROPY_INCLUDED_CXD56_COMMON_HAL_BUSIO_SPI_H
 
-#endif  // MICROPY_INCLUDED_SPRESENSE_INTERNAL_FLASH_ROOT_POINTERS_H
+#include <nuttx/spi/spi.h>
+
+#include "py/obj.h"
+
+#include "common-hal/microcontroller/Pin.h"
+
+typedef struct {
+    mp_obj_base_t base;
+    struct spi_dev_s* spi_dev;
+    uint32_t frequency;
+    uint8_t phase;
+    uint8_t polarity;
+    uint8_t bits;
+    bool has_lock;
+    const mcu_pin_obj_t *clock_pin;
+    const mcu_pin_obj_t *mosi_pin;
+    const mcu_pin_obj_t *miso_pin;
+} busio_spi_obj_t;
+
+#endif // MICROPY_INCLUDED_CXD56_COMMON_HAL_BUSIO_SPI_H
