+++ ports/cxd56/common-hal/busio/UART.h
@@ -24,16 +24,20 @@
  * THE SOFTWARE.
  */
 
-#ifndef MICROPY_INCLUDED_SPRESENSE_COMMON_HAL_PULSEIO_PULSEOUT_H
-#define MICROPY_INCLUDED_SPRESENSE_COMMON_HAL_PULSEIO_PULSEOUT_H
+#ifndef MICROPY_INCLUDED_CXD56_COMMON_HAL_BUSIO_UART_H
+#define MICROPY_INCLUDED_CXD56_COMMON_HAL_BUSIO_UART_H
 
 #include "py/obj.h"
 
+#include "common-hal/microcontroller/Pin.h"
+
 typedef struct {
     mp_obj_base_t base;
-    uint8_t pwm_num;
-} pulseio_pulseout_obj_t;
+    int uart_fd;
+    const mcu_pin_obj_t *tx_pin;
+    const mcu_pin_obj_t *rx_pin;
+    uint32_t baudrate;
+    uint32_t timeout;
+} busio_uart_obj_t;
 
-void pulseout_reset(void);
-
-#endif // MICROPY_INCLUDED_SPRESENSE_COMMON_HAL_PULSEIO_PULSEOUT_H
+#endif // MICROPY_INCLUDED_CXD56_COMMON_HAL_BUSIO_UART_H
