+++ ports/cxd56/common-hal/pulseio/PWMOut.h
@@ -24,10 +24,25 @@
  * THE SOFTWARE.
  */
 
-#ifndef MICROPY_INCLUDED_SPRESENSE_COMMON_HAL_BUSIO_ONEWIRE_H
-#define MICROPY_INCLUDED_SPRESENSE_COMMON_HAL_BUSIO_ONEWIRE_H
+#ifndef MICROPY_INCLUDED_CXD56_COMMON_HAL_PULSEIO_PWMOUT_H
+#define MICROPY_INCLUDED_CXD56_COMMON_HAL_PULSEIO_PWMOUT_H
 
-// Use bitbangio.
-#include "shared-module/busio/OneWire.h"
+#include <nuttx/drivers/pwm.h>
 
-#endif // MICROPY_INCLUDED_SPRESENSE_COMMON_HAL_BUSIO_ONEWIRE_H
+#include "common-hal/microcontroller/Pin.h"
+
+#include "py/obj.h"
+
+typedef struct {
+    mp_obj_base_t base;
+    const mcu_pin_obj_t *pin;
+    struct pwm_info_s info;
+    bool variable_frequency;
+    uint8_t number;
+} pulseio_pwmout_obj_t;
+
+void pwmout_reset(void);
+void pwmout_start(uint8_t pwm_num);
+void pwmout_stop(uint8_t pwm_num);
+
+#endif // MICROPY_INCLUDED_CXD56_COMMON_HAL_PULSEIO_PWMOUT_H
