+++ ports/cxd56/common-hal/pulseio/PulseOut.h
@@ -24,19 +24,16 @@
  * THE SOFTWARE.
  */
 
-#ifndef MICROPY_INCLUDED_SPRESENSE_COMMON_HAL_ANALOGIO_ANALOGIN_H
-#define MICROPY_INCLUDED_SPRESENSE_COMMON_HAL_ANALOGIO_ANALOGIN_H
+#ifndef MICROPY_INCLUDED_CXD56_COMMON_HAL_PULSEIO_PULSEOUT_H
+#define MICROPY_INCLUDED_CXD56_COMMON_HAL_PULSEIO_PULSEOUT_H
 
 #include "py/obj.h"
 
-#include "common-hal/microcontroller/Pin.h"
-
 typedef struct {
     mp_obj_base_t base;
-    const mcu_pin_obj_t *pin;
-    uint8_t number;
-} analogio_analogin_obj_t;
+    uint8_t pwm_num;
+} pulseio_pulseout_obj_t;
 
-void analogin_reset(void);
+void pulseout_reset(void);
 
-#endif // MICROPY_INCLUDED_SPRESENSE_COMMON_HAL_ANALOGIO_ANALOGIN_H
+#endif // MICROPY_INCLUDED_CXD56_COMMON_HAL_PULSEIO_PULSEOUT_H
