+++ ports/cxd56/common-hal/rtc/RTC.h
@@ -24,11 +24,7 @@
  * THE SOFTWARE.
  */
 
-#ifndef MICROPY_INCLUDED_SPRESENSE_TICK_H
-#define MICROPY_INCLUDED_SPRESENSE_TICK_H
+#ifndef MICROPY_INCLUDED_CXD56_COMMON_HAL_RTC_RTC_H
+#define MICROPY_INCLUDED_CXD56_COMMON_HAL_RTC_RTC_H
 
-#include "py/mpconfig.h"
-
-extern volatile uint64_t ticks_ms;
-
-#endif  // MICROPY_INCLUDED_SPRESENSE_TICK_H
+#endif  // MICROPY_INCLUDED_CXD56_COMMON_HAL_RTC_RTC_H
