+++ ports/cxd56/mphalport.c
