+++ ports/cxd56/qstrdefsport.h
