+++ ports/cxd56/tick.c
